+++ packages/nesd/lib/ui/emulator/input/action/all_actions.dart
@@ -75,6 +75,7 @@ const saveStateActions = [
 
 const toolActions = [
   toggleDisplay,
+  toggleAudio,
   toggleTileViewer,
   toggleCartridgeInfo,
   toggleDebugger,