+++ packages/nesd/test/ui/emulator/nes_controller_database_test.dart
@@ -74,6 +74,7 @@ void main() {
     when(() => settings.rewind).thenReturn(false);
     when(() => settings.volume).thenReturn(1.0);
     when(() => settings.lowPassFilter).thenReturn(false);
+    when(() => settings.swapDutyCycles).thenReturn(false);
     when(() => settings.autoSave).thenReturn(false);
     when(() => settings.autoSaveInterval).thenReturn(5);
     when(() => settings.autoLoad).thenReturn(false);