+++ packages/nesd/test/ui/emulator/nes_controller_rewind_gate_test.dart
@@ -52,6 +52,7 @@ void main() {
     when(() => settings.rewind).thenReturn(true);
     when(() => settings.volume).thenReturn(1.0);
     when(() => settings.lowPassFilter).thenReturn(false);
+    when(() => settings.swapDutyCycles).thenReturn(false);
     when(() => settings.fastForwardSpeed).thenReturn(FastForwardSpeed.x2);
     when(() => settings.turboSpeed).thenReturn(TurboSpeed.x1);
     when(() => settings.autoSave).thenReturn(false);